requirement.txt
gasgiref==3.8.1
boto3==1.34.139
botocore==1.34.139
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
diff-match-patch==20230430
dj-rest-auth==6.0.0
Django==4.1
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-import-export==4.0.9
django-oauth-toolkit==2.4.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.2.0
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-social-oauth2==3.1.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
facebook-sdk==3.1.0
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
jwcrypto==1.5.6
msgpack==1.0.8
mysqlclient==2.2.4
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
stripe==10.8.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
mysqlclient==2.2.4